feat: added match found scene
